Move the service utilities to a upper service
